Working field calculation with CUDA

+++ ElectronsAndProtons_CUDA/gpu_particles_handle_t.h
@@ -2,5 +2,5 @@
 
 struct gpu_particles_handle_t{
     particles_t * d_struct; // wskaźnik do struktury GPU
-    double* d_x,* d_y,* d_vx,* d_vy,* d_ax,* d_ay,* d_q,* d_m;
+    float * d_x,* d_y,* d_vx,* d_vy,* d_ax,* d_ay,* d_q,* d_m;
 };
